% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepJ.R
\name{deepJ}
\alias{deepJ}
\title{deepJ
Fit deepJ function}
\usage{
deepJ(model, epochs = 150, batch_size = NULL, corr = FALSE)
}
\arguments{
\item{model}{model of class deepJmodel}

\item{nLatent}{number of latent variables}
}
\description{
deepJ
Fit deepJ function
}
