% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_torch_available}
\alias{is_torch_available}
\title{is_torch_available
check whetcher torch is available}
\usage{
is_torch_available()
}
\description{
is_torch_available
check whetcher torch is available
}
