% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_varational_dense}
\alias{layer_varational_dense}
\title{add_varational_dense function}
\usage{
layer_varational_dense(model, hidden = 10L, activation = "relu",
  sd = 20, kl_weight = 0.1)
}
\arguments{
\item{model}{model}

\item{hidden}{hidden units}

\item{activation}{activation function}
}
\description{
add_varational_dense function
}
